New things!

+++ commands/Owner/restart.js
@@ -5,32 +5,49 @@ const { wait, yes, nope } = require("../../emojis.json");
 module.exports = {
   data: new SlashCommandBuilder()
     .setName("restart")
-    .setDescription("Restarts the client!"),
+    .setDescription("Restarts the client!")
+    .addBooleanOption((option) =>
+      option
+        .setName("close_process")
+        .setDescription("Should the process be closed?")
+        .setRequired(false)
+    ),
   async execute(interaction) {
     if (!admins.includes(interaction.user.id))
       return await interaction.reply({
         content: `${nope} - You are **NOT** an owner of **${client.user.username}**!`,
         ephemeral: true,
       });
-    await interaction.reply({
-      content: `Restarting... ${wait}`,
-      ephemeral: true,
-    });
-    await client.user.setActivity("Restarting...", {
-      type: "PLAYING",
-    });
-    client.destroy();
-    client.login(process.env.TOKEN);
-    require("../../events/ready");
-    require("../../client");
-    setTimeout(async () => {
-      await interaction.editReply({
-        content: `Restart Successful! ${yes}`,
+    var close_the_process =
+      interaction.options.getBoolean("close_process") || false;
+    if (close_the_process == true) {
+      await interaction.reply({
+        content:
+          "Okay, not restarting the client, just gonna close the process, see you soon then!",
         ephemeral: true,
       });
-    }, 1500);
-    await client.user.setActivity("Restart Completed!", {
-      type: "PLAYING",
-    });
+      setTimeout(async () => await process.exit(), 10000);
+    } else {
+      await interaction.reply({
+        content: `Restarting... ${wait}`,
+        ephemeral: true,
+      });
+      await client.user.setActivity("Restarting...", {
+        type: "PLAYING",
+      });
+      client.destroy();
+      client.login(process.env.TOKEN);
+      require("../../events/ready");
+      require("../../client");
+      setTimeout(async () => {
+        await interaction.editReply({
+          content: `Restart Successful! ${yes}`,
+          ephemeral: true,
+        });
+      }, 1500);
+      await client.user.setActivity("Restart Completed!", {
+        type: "PLAYING",
+      });
+    }
   },
 };
